+++ lib/djvUI/ListButton.h
@@ -54,9 +54,9 @@ namespace djv
                 void setFontFace(const std::string&);
                 void setFontSizeRole(MetricsRole);
 
-                int getElide() const;
+                size_t getElide() const;
 
-                void setElide(int);
+                void setElide(size_t);
 
                 const Layout::Margin& getInsideMargin() const;
